+++ src/audio/audio_config.h
@@ -3,7 +3,7 @@
 #include <stdlib.h>
 
 typedef struct AudioConfig {
-  int     rate;
+  size_t  rate;
   size_t  blocks;
   size_t  chan;
 } AudioConfig;
